# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Country Decision Atlas
  version: 0.1.0
paths:
...
  /api/v1/countries/compare:
    get:
      tags:
      - countries
      - cii
      summary: Compare Countries Cii
      operationId: compare_countries_cii_api_v1_countries_compare_get
      parameters:
      - name: countries
        in: query
        required: true
        schema:
          type: string
          description: Comma-separated country slugs, exactly 2 for MVP
          title: Countries
        description: Comma-separated country slugs, exactly 2 for MVP
      - name: scenario
        in: query
        required: true
        schema:
          type: string
          description: Scenario slug
          title: Scenario
        description: Scenario slug
      - name: persona
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Persona
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CiiCountryComparisonResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/countries/matrix:
    get:
      tags:
      - countries
      - cii
      summary: Get Countries Matrix
      operationId: get_countries_matrix_api_v1_countries_matrix_get
      parameters:
      - name: countries
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Countries
      - name: scenarios
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Scenarios
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareMatrixResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/countries:
    get:
      tags:
      - countries
      summary: Read Countries
      operationId: read_countries_api_v1_countries_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/countries/{country_slug}:
    get:
      tags:
      - countries
      summary: Read Country
      operationId: read_country_api_v1_countries__country_slug__get
      parameters:
      - name: country_slug
        in: path
        required: true
        schema:
          type: string
          title: Country Slug
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/countries/{country_id}/profile:
    get:
      tags:
      - countries
      summary: Read Country Profile
      operationId: read_country_profile_api_v1_countries__country_id__profile_get
      parameters:
      - name: country_id
        in: path
        required: true
        schema:
          type: string
          title: Country Id
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryProfileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/countries/{country_id}/scores:
    get:
      tags:
      - countries
      summary: Read Country Scores
      operationId: read_country_scores_api_v1_countries__country_id__scores_get
      parameters:
      - name: country_id
        in: path
        required: true
        schema:
          type: string
          title: Country Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryScoreListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/countries/{country_slug}/card:
    get:
      tags:
      - countries
      summary: Read Country Card
      operationId: read_country_card_api_v1_countries__country_slug__card_get
      parameters:
      - name: country_slug
        in: path
        required: true
        schema:
          type: string
          title: Country Slug
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryReadModelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/countries/{country_slug}/cii:
    get:
      tags:
      - countries
      summary: Read Country Cii
      operationId: read_country_cii_api_v1_countries__country_slug__cii_get
      parameters:
      - name: country_slug
        in: path
        required: true
        schema:
          type: string
          title: Country Slug
      - name: version
        in: query
        required: false
        schema:
          type: string
          pattern: ^v\d+\.\d+$
          default: v1.0
          title: Version
      - name: scenario
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Scenario
      - name: persona
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Persona
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryReadModelCii'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/countries/{country_slug}/sources:
    get:
      tags:
      - countries
      summary: Read Country Sources
      operationId: read_country_sources_api_v1_countries__country_slug__sources_get
      parameters:
      - name: country_slug
        in: path
        required: true
        schema:
          type: string
          title: Country Slug
      - name: source_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Type
      - name: language
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
      - name: confidence
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - low
            - medium
            - high
            type: string
          - type: 'null'
          title: Confidence
      - name: status
        in: query
        required: false
        schema:
          enum:
          - published
          - archived
          type: string
          default: published
          title: Status
      - name: sort
        in: query
        required: false
        schema:
          enum:
          - title
          - created_at
          - published_at
          - last_checked_at
          - confidence
          type: string
          default: title
          title: Sort
      - name: order
        in: query
        required: false
        schema:
          enum:
          - asc
          - desc
          type: string
          default: asc
          title: Order
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceListWithLocaleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/scenarios/{slug}/countries:
    get:
      tags:
      - scenarios
      summary: Read Scenario Countries
      operationId: read_scenario_countries_api_v1_scenarios__slug__countries_get
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionCountryScoreListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/countries/{country_slug}/profile:
    patch:
      tags:
      - admin
      summary: Admin Patch Country Profile
      operationId: admin_patch_country_profile_api_v1_admin_countries__country_slug__profile_patch
      security:
      - AdminTokenAuth: []
      parameters:
      - name: country_slug
        in: path
        required: true
        schema:
          type: string
          title: Country Slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountryProfilePatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCountryProfileResponse'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentValidationError'
          description: Unprocessable Content
...
components:
  schemas:
    AdminCountryProfileResponse:
      properties:
        item:
          additionalProperties: true
          type: object
          title: Item
        audit:
          anyOf:
          - $ref: '#/components/schemas/AuditEvent'
          - type: 'null'
      type: object
      required:
      - item
      title: AdminCountryProfileResponse
...
    CiiCountryComparisonResponse:
      properties:
        scenario:
          $ref: '#/components/schemas/ComparedScenario'
        locale:
          $ref: '#/components/schemas/LocaleResolution'
        countries:
          items:
            $ref: '#/components/schemas/ComparedCountry'
          type: array
          title: Countries
        metrics:
          items:
            $ref: '#/components/schemas/ComparedMetric'
          type: array
          title: Metrics
        formula_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Formula Version
        aggregation_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Aggregation Method
        weights_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Weights Version
        quality_warnings:
          items:
            type: string
          type: array
          title: Quality Warnings
        applied_persona:
          anyOf:
          - $ref: '#/components/schemas/Persona'
          - type: 'null'
        persona_weight_profile:
          anyOf:
          - $ref: '#/components/schemas/PersonaWeightProfile'
          - type: 'null'
      type: object
      required:
      - scenario
      - locale
      title: CiiCountryComparisonResponse
    CompareMatrixResponse:
      properties:
        locale:
          $ref: '#/components/schemas/LocaleResolution'
        countries:
          items:
            $ref: '#/components/schemas/MatrixCountry'
          type: array
          title: Countries
        scenarios:
          items:
            $ref: '#/components/schemas/MatrixScenario'
          type: array
          title: Scenarios
        cells:
          items:
            $ref: '#/components/schemas/MatrixCell'
          type: array
          title: Cells
        formula_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Formula Version
        aggregation_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Aggregation Method
        weights_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Weights Version
        quality_warnings:
          items:
            type: string
          type: array
          title: Quality Warnings
      type: object
      required:
      - locale
      title: CompareMatrixResponse
    ComparedCountry:
      properties:
        slug:
          type: string
          title: Slug
        name:
          type: string
          title: Name
        iso2:
          anyOf:
          - type: string
          - type: 'null'
          title: Iso2
        cii_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Cii Score
        cii_confidence:
          anyOf:
          - type: string
          - type: 'null'
          title: Cii Confidence
        country_drift:
          anyOf:
          - type: number
          - type: 'null'
          title: Country Drift
      type: object
      required:
      - slug
      - name
      title: ComparedCountry
...
    ComparedScenario:
      properties:
        slug:
          type: string
          title: Slug
        title:
          type: string
          title: Title
      type: object
      required:
      - slug
      - title
      title: ComparedScenario
    ContentValidationError:
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
      type: object
      required:
      - error
      title: ContentValidationError
...
    CountryListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Country'
          type: array
          title: Items
        pagination:
          $ref: '#/components/schemas/Pagination'
        locale:
          $ref: '#/components/schemas/LocaleResolution'
      type: object
      required:
      - items
      - pagination
      - locale
      title: CountryListResponse
...
    CountryProfilePatch:
      properties:
        locale:
          type: string
          title: Locale
          default: en
        executive_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Executive Summary
        migration_overview:
          anyOf:
          - type: string
          - type: 'null'
          title: Migration Overview
        tax_overview:
          anyOf:
          - type: string
          - type: 'null'
          title: Tax Overview
        cost_of_living_overview:
          anyOf:
          - type: string
          - type: 'null'
          title: Cost Of Living Overview
        business_overview:
          anyOf:
          - type: string
          - type: 'null'
          title: Business Overview
        safety_overview:
          anyOf:
          - type: string
          - type: 'null'
          title: Safety Overview
        legal_signals_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Legal Signals Summary
        risk_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk Summary
        source_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Summary
        status:
          anyOf:
          - $ref: '#/components/schemas/PublicationStatus'
          - type: 'null'
      type: object
      title: CountryProfilePatch
    CountryProfileResponse:
      properties:
        item:
          $ref: '#/components/schemas/CountryProfile'
        locale:
          $ref: '#/components/schemas/LocaleResolution'
      type: object
      required:
      - item
      - locale
      title: CountryProfileResponse
    CountryReadModelCii:
      properties:
        overall_score:
          type: number
          title: Overall Score
        confidence:
          type: string
          title: Confidence
        drift:
          anyOf:
          - type: number
          - type: 'null'
          title: Drift
        version:
          type: string
          title: Version
        formula_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Formula Version
        aggregation_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Aggregation Method
        quality_warnings:
          items:
            type: string
          type: array
          title: Quality Warnings
        calculated_at:
          type: string
          format: date-time
          title: Calculated At
        metrics:
          items:
            $ref: '#/components/schemas/CountryReadModelCiiMetric'
          type: array
          title: Metrics
        applied_persona:
          anyOf:
          - $ref: '#/components/schemas/Persona'
          - type: 'null'
        persona_weight_profile:
          anyOf:
          - $ref: '#/components/schemas/PersonaWeightProfile'
          - type: 'null'
      type: object
      required:
      - overall_score
      - confidence
      - version
      - calculated_at
      title: CountryReadModelCii
...
    CountryReadModelResponse:
      properties:
        country:
          $ref: '#/components/schemas/CountryReadModelCountry'
        profile:
          anyOf:
          - $ref: '#/components/schemas/CountryReadModelProfile'
          - type: 'null'
        scores:
          items:
            $ref: '#/components/schemas/CountryReadModelScore'
          type: array
          title: Scores
        legal_signals:
          items:
            $ref: '#/components/schemas/CountryReadModelLegalSignal'
          type: array
          title: Legal Signals
        sources:
          items:
            $ref: '#/components/schemas/CountryReadModelSource'
          type: array
          title: Sources
        evidence_summary:
          $ref: '#/components/schemas/CountryReadModelEvidenceSummary'
        user_stories_summary:
          $ref: '#/components/schemas/CountryReadModelUserStoriesSummary'
        cii:
          anyOf:
          - $ref: '#/components/schemas/CountryReadModelCii'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/CountryReadModelMeta'
        locale:
          $ref: '#/components/schemas/LocaleResolution'
      type: object
      required:
      - country
      - profile
      - scores
      - legal_signals
      - sources
      - evidence_summary
      - user_stories_summary
      - meta
      - locale
      title: CountryReadModelResponse
...
    CountryResponse:
      properties:
        item:
          $ref: '#/components/schemas/Country'
        locale:
          $ref: '#/components/schemas/LocaleResolution'
      type: object
      required:
      - item
      - locale
      title: CountryResponse
...
    CountryScoreListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CountryScore'
          type: array
          title: Items
        pagination:
          $ref: '#/components/schemas/Pagination'
        locale:
          $ref: '#/components/schemas/LocaleResolution'
      type: object
      required:
      - items
      - pagination
      - locale
      title: CountryScoreListResponse
...
    DecisionCountryScoreListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/DecisionCountryScore'
          type: array
          title: Items
        locale:
          $ref: '#/components/schemas/LocaleResolution'
      type: object
      required:
      - items
      - locale
      title: DecisionCountryScoreListResponse
...
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
      type: object
      required:
      - error
      title: ErrorResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LocaleCode:
      type: string
      enum:
      - en
      - ru
      title: LocaleCode
    LocaleResolution:
      properties:
        requested_locale:
          $ref: '#/components/schemas/LocaleCode'
        resolved_locale:
          $ref: '#/components/schemas/LocaleCode'
        translation_status:
          $ref: '#/components/schemas/TranslationStatus'
      type: object
      required:
      - requested_locale
      - resolved_locale
      - translation_status
      title: LocaleResolution
...
    SourceListWithLocaleResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Source'
          type: array
          title: Items
        pagination:
          $ref: '#/components/schemas/Pagination'
        sort:
          anyOf:
          - $ref: '#/components/schemas/SortMeta'
          - type: 'null'
        locale:
          $ref: '#/components/schemas/LocaleResolution'
      type: object
      required:
      - items
      - pagination
      - locale
      title: SourceListWithLocaleResponse
...
  securitySchemes:
    AdminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token